This works better for now, but relinks. Makefile war is not over

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: juramos <juramos@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/21 10:35:27 by juramos           #+#    #+#             */
-/*   Updated: 2023/11/28 12:33:24 by juramos          ###   ########.fr       */
+/*   Updated: 2023/11/28 12:48:52 by juramos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 # include <stdio.h>
 # include "libft/libft.h"
 
-int	ft_printf(char const *format, ...);
-int	ft_write_s(va_list args);
+int		ft_printf(char const *format, ...);
+size_t	ft_write_s(va_list args);
 
 #endif
